+++ crates/runtime/src/flight/middleware/request_context.rs
@@ -1,0 +1,94 @@
+/*
+Copyright 2024-2025 The Spice.ai OSS Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+     https://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+use std::{
+    future::Future,
+    pin::Pin,
+    sync::Arc,
+    task::{Context, Poll},
+};
+
+use crate::request::{Protocol, RequestContext};
+use app::App;
+
+use runtime_auth::AuthRequestContext;
+use tower::{Layer, Service};
+
+/// Extracts the request context from the HTTP headers and adds it to the task-local context.
+#[derive(Clone)]
+pub struct RequestContextLayer {
+    app: Option<Arc<App>>,
+}
+
+impl RequestContextLayer {
+    #[must_use]
+    pub fn new(app: Option<Arc<App>>) -> Self {
+        Self { app }
+    }
+}
+
+impl<S> Layer<S> for RequestContextLayer {
+    type Service = RequestContextMiddleware<S>;
+
+    fn layer(&self, inner: S) -> Self::Service {
+        RequestContextMiddleware {
+            inner,
+            app: self.app.clone(),
+        }
+    }
+}
+
+#[derive(Clone)]
+pub struct RequestContextMiddleware<S> {
+    inner: S,
+    app: Option<Arc<App>>,
+}
+
+impl<S, ReqBody, ResBody> Service<http::Request<ReqBody>> for RequestContextMiddleware<S>
+where
+    S: Service<http::Request<ReqBody>, Response = http::Response<ResBody>> + Clone + Send + 'static,
+    S::Future: Send + 'static,
+    ResBody: Default,
+    ReqBody: Send + 'static,
+{
+    type Response = S::Response;
+    type Error = S::Error;
+    type Future = Pin<Box<dyn Future<Output = Result<S::Response, S::Error>> + Send>>;
+
+    fn poll_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        self.inner.poll_ready(cx)
+    }
+
+    fn call(&mut self, mut req: http::Request<ReqBody>) -> Self::Future {
+        let clone = self.inner.clone();
+        let mut inner = std::mem::replace(&mut self.inner, clone);
+
+        let headers = req.headers();
+        let request_context = Arc::new(
+            RequestContext::builder(Protocol::Flight)
+                .with_app_opt(self.app.clone())
+                .from_headers(headers)
+                .build(),
+        );
+
+        req.extensions_mut()
+            .insert::<Arc<dyn AuthRequestContext + Send + Sync>>(
+                Arc::clone(&request_context) as Arc<dyn AuthRequestContext + Send + Sync>
+            );
+
+        Box::pin(async move { request_context.scope(inner.call(req)).await })
+    }
+}
